fix slog_014_pos test
XXX I might have lost the essence of this change during a rebase.
Look at commit 31073a1db4a6834638fc2fa47d74e8c2d556e9d2 in
old/zfs-tests-fixes.

First, the tested configurations are redundant and the scrub speed seems
to be fast enough that the pools are not seen in the DEGRADED state.
Second, no status is actually reported on the 'logs' line.  Status of
each slog device is reported individually,

Also, increase the size of the test devices.  These days more space is
reserved in the pool, so less space is available for data.
In particuar, the creation of the filler file used to fail.